+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/settings/viewmodel/SettingsViewModel.kt
@@ -18,6 +18,7 @@ import co.electriccoin.zcash.ui.common.model.ZashiAccount
 import co.electriccoin.zcash.ui.common.provider.GetVersionInfoProvider
 import co.electriccoin.zcash.ui.common.usecase.IsCoinbaseAvailableUseCase
 import co.electriccoin.zcash.ui.common.usecase.IsFlexaAvailableUseCase
+import co.electriccoin.zcash.ui.common.usecase.NavigateToAddressBookUseCase
 import co.electriccoin.zcash.ui.common.usecase.ObserveConfigurationUseCase
 import co.electriccoin.zcash.ui.common.usecase.ObserveSelectedWalletAccountUseCase
 import co.electriccoin.zcash.ui.common.usecase.RescanBlockchainUseCase
@@ -51,6 +52,7 @@ class SettingsViewModel(
     private val getVersionInfo: GetVersionInfoProvider,
     private val rescanBlockchain: RescanBlockchainUseCase,
     private val navigationRouter: NavigationRouter,
+    private val navigateToAddressBook: NavigateToAddressBookUseCase
 ) : ViewModel() {
     private val versionInfo by lazy { getVersionInfo() }
 
@@ -230,7 +232,10 @@ class SettingsViewModel(
 
     private fun onSendUsFeedbackClick() = navigationRouter.forward(SUPPORT)
 
-    private fun onAddressBookClick() = navigationRouter.forward(AddressBookArgs(AddressBookArgs.DEFAULT))
+    private fun onAddressBookClick() =
+        viewModelScope.launch {
+            navigateToAddressBook(AddressBookArgs.DEFAULT)
+        }
 
     private fun onWhatsNewClick() = navigationRouter.forward(WHATS_NEW)