+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 ## [Unreleased]
 
 ### Changed
+- Send Confirmation & Send Progress screens have been refactored
 - ZXing QR codes scanning library has been replaced with a more recent MLkit Barcodes scanning library, which gives 
   us better results in testing